requirements.txt
# ── Core Data Science ─────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn>=1.3.0
xgboost>=2.0.0

# ── NLP ───────────────────────────────────────────────────────────────────────
transformers>=4.33.0
torch>=2.0.0
sentencepiece>=0.1.99
spacy>=3.6.0
# After install: python -m spacy download en_core_web_sm

# ── Computer Vision (optional) ────────────────────────────────────────────────
# torchvision>=0.15.0
# timm>=0.9.7
# opencv-python>=4.8.0

# ── Data Collection ───────────────────────────────────────────────────────────
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
newsapi-python>=0.2.7

# ── Deployment & UI ───────────────────────────────────────────────────────────
streamlit>=1.28.0
plotly>=5.17.0
Pillow>=10.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0
pyyaml>=6.0.0
